Homework, Week 4
I did both projects this week.  Since I worked more on PyCity I will submit that one as my primary.  It was a challenging project, as summarizing the percentage of passing marks was tricky.   I also use a combination of Groupby and Pivot tables in order to finish the projected.  I find pivot tables to be somewhat easier to manage than Groupby.  

I can't help but think that the way I solved the pass percentage problem was not as elegant as I wanted; I would have preferred to include a counting function for a condition in the Aggregators for the Groupby, but I could not figure out the syntax.  

Additionally, I formatted most of the text and included a couple of charts to finish of the project.